--- src/main/java/edu/netcracker/messenger/GeneratePassword.java
@@ -1,13 +1,0 @@
-package edu.netcracker.messenger;
-
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-public class GeneratePassword {
-//    public static void main(String[] args) {
-//        BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
-//        String pass = "kirill123";
-//        String encPass = encoder.encode(pass);
-//
-//        System.out.println(encPass);
-//    }
-}
