+++ src/main/java/edu/netcracker/messenger/MessengerApplication.java
@@ -2,9 +2,7 @@ package edu.netcracker.messenger;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-//@SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
 @SpringBootApplication
 public class MessengerApplication {
