--- src/main/java/edu/netcracker/messenger/config/MvcConfiguration.java
@@ -1,12 +1,0 @@
-package edu.netcracker.messenger.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class MvcConfiguration implements WebMvcConfigurer {
-//    public void addViewControllers(ViewControllerRegistry registry) {
-//        registry.addViewController("/login").setViewName("login");
-//    }
-}
